# api/docs/swagger.yaml
definitions:
  model.CheckStudentTestReq:
    properties:
      answers:
        items:
          $ref: '#/definitions/model.QuestionAnswer'
        type: array
      day:
        type: string
      student_id:
        type: string
    type: object
...
  model.CreateTemplateReq:
    properties:
      day:
        type: string
      student_id:
        type: string
    type: object
  model.Error:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  model.QuestionAnswer:
    properties:
      answer:
        type: string
      number:
        type: integer
    type: object
...
  model.Result:
    properties:
      correct:
        type: integer
      incorrect:
        type: integer
      percent:
        type: number
    type: object
  model.Status:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /templates/check:
    post:
      consumes:
      - application/json
      description: Ushbu endpoint talabaga tegishli test javoblarini tekshirish uchun
        ishlatiladi
      parameters:
      - description: Talabaning test javoblarini tekshirish uchun ma'lumot
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CheckStudentTestReq'
      produces:
      - application/json
      responses:
        "200":
          description: Test natijalari muvaffaqiyatli tekshirildi
          schema:
            $ref: '#/definitions/model.Result'
        "400":
          description: Noto'g'ri ma'lumot kiritildi
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Talabaning test natijalarini tekshirish
      tags:
      - Templates
  /templates/create:
    post:
      consumes:
      - application/json
      description: Yangi test template yaratish
      parameters:
      - description: Template ma'lumotlari
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/model.CreateTemplateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Muvaffaqiyatli yaratildi
          schema:
            $ref: '#/definitions/model.Status'
        "400":
          description: Noto'g'ri ma'lumot kiritildi
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Serverda xatolik
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Template yaratish
      tags:
      - Templates
  /templates/get:
    get:
      consumes:
      - application/json
      description: Student ID va sana bo‘yicha shablonni yuklab olish
      parameters:
      - description: Talaba ID
        in: query
        name: student_id
        required: true
        type: string
      - description: Kun (YYYY-MM-DD format)
        in: query
        name: day
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Shablon fayli
          schema:
            type: file
        "400":
          description: Xato so‘rov
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Server xatosi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Student shablonini yuklab olish
      tags:
      - Templates
securityDefinitions:
  BearerAuth:
    description: 'Tokenni faqat o‘zi yozing, masalan: eyJhbGciOiJIUzI1NiIs..'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"